Actually handle unicode properly
A problem with b-encoding is that byte strings are used both
as text strings and as buckets for binary data. That becomes
a problem when you're trying to parse them as strings.

I think this change is the ideal compromise: if a string is pure ASCII,
it's decoded as a string. If not, it's left totally alone. That leaves
the unicode-decoding totally up to the consumer, which is appropriate,
and no bytes are ever modified.

Strings to get encoded as UTF-8, but that's common practice.